create new baseViewModel to move functions
